app.rb: Remove Sinatra::Base application wrapper
Don't wrap our routes/controllers in a `Sinatra::Base` class to keep
things simple. Was doing this because I thought I needed to in order to
be able to tell Capybara where the application was but looking at the
Sinatra testing page (http://www.sinatrarb.com/testing.html), they
provide an example for Capybara that details this
`Sinatra::Application.new` call that I can make instead. This frees us
up to use the simplified application definition.